Improved the structure of inc operation visualization in Eva. Also fixed a bug ReplActor was causing for :quit by keeping its thread alive.
